Refactor cache dependency handling in repositories
Updated BaseRepository to use ICacheDependencyBuilder for creating cache dependencies. Introduced CacheDependencyBuilder class and ICacheDependencyBuilder interface. Modified ContentTypeRepository and PageTypeRepository to pass ICacheDependencyBuilder to BaseRepository. Removed static methods from BaseRepository, moving their functionality to CacheDependencyBuilder.
